refactor write_hdf5_with_json_support()
