+++ pyiron_base/storage/hdfio.py
@@ -17,7 +17,12 @@ import sys
 from typing import Union, Optional, Any, Tuple
 
 from pyiron_base.utils.deprecate import deprecate
-from pyiron_base.storage.helper_functions import read_hdf5, write_hdf5
+from pyiron_base.storage.helper_functions import (
+    open_hdf5,
+    read_hdf5,
+    write_hdf5_with_json_support,
+    _is_ragged_in_1st_dim_only,
+)
 from pyiron_base.interfaces.has_groups import HasGroups
 from pyiron_base.state import state
 from pyiron_base.jobs.dynamic import JOB_DYN_DICT, class_constructor
@@ -35,52 +40,6 @@ __maintainer__ = "Jan Janssen"
 __email__ = "janssen@mpie.de"
 __status__ = "production"
 __date__ = "Sep 1, 2017"
-
-
-def _is_ragged_in_1st_dim_only(value: Union[np.ndarray, list]) -> bool:
-    """
-    Checks whether array or list of lists is ragged in the first dimension.
-
-    That means all other dimensions (except the first one) still have to match.
-
-    Args:
-        value (ndarray/list): array to check
-
-    Returns:
-        bool: True if elements of value are not all of the same shape
-    """
-    if isinstance(value, np.ndarray) and value.dtype != np.dtype("O"):
-        return False
-    else:
-
-        def extract_dims(v):
-            with warnings.catch_warnings():
-                warnings.simplefilter("ignore")
-                s = np.shape(v)
-            return s[0], s[1:]
-
-        dim1, dim_other = zip(*map(extract_dims, value))
-        return len(set(dim1)) > 1 and len(set(dim_other)) == 1
-
-
-def _check_json_conversion(value):
-    use_json = True
-    if (
-            isinstance(value, (list, np.ndarray))
-            and len(value) > 0
-            and isinstance(value[0], (list, np.ndarray))
-            and len(value[0]) > 0
-            and not isinstance(value[0][0], str)
-            and _is_ragged_in_1st_dim_only(value)
-    ):
-        # if the sub-arrays in value all share shape[1:], h5io comes up with a more efficient storage format than
-        # just writing a dataset for each element, by concatenating along the first axis and storing the indices
-        # where to break the concatenated array again
-        value = np.array([np.asarray(v) for v in value], dtype=object)
-        use_json = False
-    elif isinstance(value, tuple):
-        value = list(value)
-    return value, use_json
 
 
 # for historic reasons we write str(class) into the HDF 'TYPE' field of objects, so we need to parse this back out
@@ -182,15 +141,6 @@ def _to_object(hdf, class_name=None, **kwargs):
     return obj
 
 
-def open_hdf5(filename, mode="r", swmr=False):
-    if swmr and mode != "r":
-        store = h5py.File(filename, mode=mode, libver="latest")
-        store.swmr = True
-        return store
-    else:
-        return h5py.File(filename, mode=mode, libver="latest", swmr=swmr)
-
-
 class FileHDFio(HasGroups, MutableMapping):
     """
     Class that provides all info to access a h5 file. This class is based on h5io.py, which allows to
@@ -269,7 +219,7 @@ class FileHDFio(HasGroups, MutableMapping):
                 # underlying file once, this reduces the number of file opens in the most-likely case from 2 to 1 (1 to
                 # check whether the data is there and 1 to read it) and increases in the worst case from 1 to 2 (1 to
                 # try to read it here and one more time to verify it's not a group below).
-                return read_hdf5(self.file_name, title=self._get_h5_path(item))
+                return read_hdf5(self.file_name, title=self.get_h5_path(item))
             except (ValueError, OSError, RuntimeError, NotImplementedError):
                 # h5io couldn't find a dataset with name item, but there still might be a group with that name, which we
                 # check in the rest of the method
@@ -362,14 +312,8 @@ class FileHDFio(HasGroups, MutableMapping):
         ):
             value.to_hdf(self, key)
             return
-
-        value, use_json = _check_json_conversion(value=value)
-        write_hdf5(
-            self.file_name,
-            value,
-            title=self._get_h5_path(key),
-            overwrite="update",
-            use_json=use_json,
+        write_hdf5_with_json_support(
+            value=value, path=self.get_h5_path(key), file_handle=self.file_name
         )
 
     def __delitem__(self, key):
@@ -382,7 +326,7 @@ class FileHDFio(HasGroups, MutableMapping):
         if self.file_exists:
             try:
                 with open_hdf5(self.file_name, mode="a") as store:
-                    del store[self._get_h5_path(key)]
+                    del store[self.get_h5_path(key)]
             except (AttributeError, KeyError):
                 pass
 
@@ -638,7 +582,7 @@ class FileHDFio(HasGroups, MutableMapping):
         Returns:
             FileHDFio: FileHDFio object pointing to the new group
         """
-        full_name = self._get_h5_path(name)
+        full_name = self.get_h5_path(name)
         with open_hdf5(self.file_name, mode="a") as h:
             try:
                 h.create_group(full_name, track_order=track_order)
@@ -677,7 +621,7 @@ class FileHDFio(HasGroups, MutableMapping):
         if h5_rel_path.strip() == ".":
             h5_rel_path = ""
         if h5_rel_path.strip() != "":
-            new_h5_path.h5_path = self._get_h5_path(h5_rel_path)
+            new_h5_path.h5_path = self.get_h5_path(h5_rel_path)
         new_h5_path.history.append(h5_rel_path)
 
         return new_h5_path
@@ -1008,7 +952,7 @@ class FileHDFio(HasGroups, MutableMapping):
         Returns:
             dict, list, float, int: data or data object
         """
-        return read_hdf5(self.file_name, title=self._get_h5_path(item))
+        return read_hdf5(self.file_name, title=self.get_h5_path(item))
 
     # def _open_store(self, mode="r"):
     #     """
@@ -1044,7 +988,7 @@ class FileHDFio(HasGroups, MutableMapping):
 
         return Project(path=self.file_path)
 
-    def _get_h5_path(self, name):
+    def get_h5_path(self, name):
         """
         Internal function to combine the current h5_path with the relative path
