+++ pyiron_base/storage/helper_functions.py
@@ -1,5 +1,17 @@
+import numpy as np
 import h5io
+import h5py
 from pyiron_base.utils.error import retry
+from typing import Union
+
+
+def open_hdf5(filename, mode="r", swmr=False):
+    if swmr and mode != "r":
+        store = h5py.File(filename, mode=mode, libver="latest")
+        store.swmr = True
+        return store
+    else:
+        return h5py.File(filename, mode=mode, libver="latest", swmr=swmr)
 
 
 def read_hdf5(fname, title="h5io", slash="ignore"):
@@ -40,3 +52,60 @@ def write_hdf5(
         at_most=10,
         delay=1,
     )
+
+
+def write_hdf5_with_json_support(value, path, file_handle):
+    value, use_json = _check_json_conversion(value=value)
+    write_hdf5(
+        file_handle,
+        value,
+        title=path,
+        overwrite="update",
+        use_json=use_json,
+    )
+
+
+def _check_json_conversion(value):
+    use_json = True
+    if (
+        isinstance(value, (list, np.ndarray))
+        and len(value) > 0
+        and isinstance(value[0], (list, np.ndarray))
+        and len(value[0]) > 0
+        and not isinstance(value[0][0], str)
+        and _is_ragged_in_1st_dim_only(value)
+    ):
+        # if the sub-arrays in value all share shape[1:], h5io comes up with a more efficient storage format than
+        # just writing a dataset for each element, by concatenating along the first axis and storing the indices
+        # where to break the concatenated array again
+        value = np.array([np.asarray(v) for v in value], dtype=object)
+        use_json = False
+    elif isinstance(value, tuple):
+        value = list(value)
+    return value, use_json
+
+
+def _is_ragged_in_1st_dim_only(value: Union[np.ndarray, list]) -> bool:
+    """
+    Checks whether array or list of lists is ragged in the first dimension.
+
+    That means all other dimensions (except the first one) still have to match.
+
+    Args:
+        value (ndarray/list): array to check
+
+    Returns:
+        bool: True if elements of value are not all of the same shape
+    """
+    if isinstance(value, np.ndarray) and value.dtype != np.dtype("O"):
+        return False
+    else:
+
+        def extract_dims(v):
+            with warnings.catch_warnings():
+                warnings.simplefilter("ignore")
+                s = np.shape(v)
+            return s[0], s[1:]
+
+        dim1, dim_other = zip(*map(extract_dims, value))
+        return len(set(dim1)) > 1 and len(set(dim_other)) == 1
